#1028 Deduplicate two files for gcc generic puller, combined into one, now can run gcc.puller_functions.py directly to pull a layer
